delete components that we won't implement
