--- frontend/src/components/LandingLoggedIn.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-import RecommendationForm from "./RecommendationForm";
-
-const LandingLoggedIn = () => {
-  return (
-    <>
-      <h1>Hello Anna!</h1>
-      <h2>Let's find your perfect wine</h2>
-      <RecommendationForm />
-    </>
-  );
-};
-
-export default LandingLoggedIn;
-
-// The landing page after the user has logged in
